requirements.txt
# TCC - Previsão HH (itens não programados)
# Python 3.9+

# Dados e pipeline
pandas>=1.5.0
numpy>=1.23.0
openpyxl>=3.0.0
PyYAML>=6.0

# Modelo e métricas
tensorflow>=2.12.0
scikit-learn>=1.2.0
joblib>=1.2.0

# API e front (produção)
flask>=2.3.0

# Opcional (notebooks e visualização)
# jupyter>=1.0.0
# seaborn>=0.12.0
# matplotlib>=3.6.0
